+++ backend/libraries/chat_events/src/chat_event_internal.rs
@@ -19,7 +19,62 @@ use types::{
 };
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
+#[serde(from = "ChatEventInternalPrevious")]
 pub enum ChatEventInternal {
+    #[serde(rename = "m")]
+    Message(Box<MessageInternal>),
+    #[serde(rename = "dcc")]
+    DirectChatCreated(DirectChatCreated),
+    #[serde(rename = "gcc")]
+    GroupChatCreated(Box<GroupCreated>),
+    #[serde(rename = "nc")]
+    GroupNameChanged(Box<GroupNameChanged>),
+    #[serde(rename = "dc")]
+    GroupDescriptionChanged(Box<GroupDescriptionChanged>),
+    #[serde(rename = "grc")]
+    GroupRulesChanged(Box<GroupRulesChanged>),
+    #[serde(rename = "ac")]
+    AvatarChanged(Box<AvatarChanged>),
+    #[serde(rename = "ma")]
+    ParticipantsAdded(Box<MembersAdded>),
+    #[serde(rename = "mr")]
+    ParticipantsRemoved(Box<MembersRemoved>),
+    #[serde(rename = "mj")]
+    ParticipantJoined(Box<MemberJoined>),
+    #[serde(rename = "ml")]
+    ParticipantLeft(Box<MemberLeft>),
+    #[serde(rename = "rc")]
+    RoleChanged(Box<RoleChanged>),
+    #[serde(rename = "ub")]
+    UsersBlocked(Box<UsersBlocked>),
+    #[serde(rename = "uub")]
+    UsersUnblocked(Box<UsersUnblocked>),
+    #[serde(rename = "mp")]
+    MessagePinned(Box<MessagePinned>),
+    #[serde(rename = "mup")]
+    MessageUnpinned(Box<MessageUnpinned>),
+    #[serde(rename = "pc")]
+    PermissionsChanged(Box<PermissionsChanged>),
+    #[serde(rename = "vc")]
+    GroupVisibilityChanged(Box<GroupVisibilityChanged>),
+    #[serde(rename = "icc")]
+    GroupInviteCodeChanged(Box<GroupInviteCodeChanged>),
+    #[serde(rename = "fz")]
+    ChatFrozen(Box<GroupFrozen>),
+    #[serde(rename = "ufz")]
+    ChatUnfrozen(Box<GroupUnfrozen>),
+    #[serde(rename = "ttl")]
+    EventsTimeToLiveUpdated(Box<EventsTimeToLiveUpdated>),
+    #[serde(rename = "gu")]
+    GroupGateUpdated(Box<GroupGateUpdated>),
+    #[serde(rename = "ui")]
+    UsersInvited(Box<UsersInvited>),
+    #[serde(rename = "e")]
+    Empty,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub enum ChatEventInternalPrevious {
     #[serde(rename = "m")]
     Message(Box<MessageInternal>),
     #[serde(rename = "dcc")]
@@ -99,14 +154,10 @@ impl ChatEventInternal {
                 | ChatEventInternal::GroupDescriptionChanged(_)
                 | ChatEventInternal::GroupRulesChanged(_)
                 | ChatEventInternal::AvatarChanged(_)
-                | ChatEventInternal::OwnershipTransferred(_)
                 | ChatEventInternal::ParticipantsAdded(_)
                 | ChatEventInternal::ParticipantsRemoved(_)
                 | ChatEventInternal::ParticipantJoined(_)
                 | ChatEventInternal::ParticipantLeft(_)
-                | ChatEventInternal::ParticipantAssumesSuperAdmin(_)
-                | ChatEventInternal::ParticipantDismissedAsSuperAdmin(_)
-                | ChatEventInternal::ParticipantRelinquishesSuperAdmin(_)
                 | ChatEventInternal::RoleChanged(_)
                 | ChatEventInternal::UsersBlocked(_)
                 | ChatEventInternal::UsersUnblocked(_)
@@ -729,6 +780,45 @@ impl ChatMetricsInternal {
             message_reminders: self.message_reminders,
             custom_type_messages: self.custom_type_messages,
             last_active: self.last_active,
+        }
+    }
+}
+
+impl From<ChatEventInternalPrevious> for ChatEventInternal {
+    fn from(value: ChatEventInternalPrevious) -> Self {
+        use ChatEventInternal as C; // Current
+        use ChatEventInternalPrevious as P; // Previous
+
+        match value {
+            P::Message(x) => C::Message(x),
+            P::DirectChatCreated(x) => C::DirectChatCreated(x),
+            P::GroupChatCreated(x) => C::GroupChatCreated(x),
+            P::GroupNameChanged(x) => C::GroupNameChanged(x),
+            P::GroupDescriptionChanged(x) => C::GroupDescriptionChanged(x),
+            P::GroupRulesChanged(x) => C::GroupRulesChanged(x),
+            P::AvatarChanged(x) => C::AvatarChanged(x),
+            P::ParticipantsAdded(x) => C::ParticipantsAdded(x),
+            P::ParticipantsRemoved(x) => C::ParticipantsRemoved(x),
+            P::ParticipantJoined(x) => C::ParticipantJoined(x),
+            P::ParticipantLeft(x) => C::ParticipantLeft(x),
+            P::RoleChanged(x) => C::RoleChanged(x),
+            P::UsersBlocked(x) => C::UsersBlocked(x),
+            P::UsersUnblocked(x) => C::UsersUnblocked(x),
+            P::MessagePinned(x) => C::MessagePinned(x),
+            P::MessageUnpinned(x) => C::MessageUnpinned(x),
+            P::PermissionsChanged(x) => C::PermissionsChanged(x),
+            P::GroupVisibilityChanged(x) => C::GroupVisibilityChanged(x),
+            P::GroupInviteCodeChanged(x) => C::GroupInviteCodeChanged(x),
+            P::ChatFrozen(x) => C::ChatFrozen(x),
+            P::ChatUnfrozen(x) => C::ChatUnfrozen(x),
+            P::EventsTimeToLiveUpdated(x) => C::EventsTimeToLiveUpdated(x),
+            P::GroupGateUpdated(x) => C::GroupGateUpdated(x),
+            P::UsersInvited(x) => C::UsersInvited(x),
+            P::OwnershipTransferred(_)
+            | P::ParticipantAssumesSuperAdmin(_)
+            | P::ParticipantDismissedAsSuperAdmin(_)
+            | P::ParticipantRelinquishesSuperAdmin(_)
+            | P::Empty => C::Empty,
         }
     }
 }