+++ backend/canisters/group/CHANGELOG.md
@@ -9,6 +9,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Refactored `claim_prize` ([#3854](https://github.com/open-chat-labs/open-chat/pull/3854))
 
+### Removed
+
+- Remove deprecated event types ([#3862](https://github.com/open-chat-labs/open-chat/pull/3862))
+
 ## [[2.0.735](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.724-group)] - 2023-06-27
 
 ### Added